ValidationError-ify `Surface` and `Swapchain`
