+++ examples/src/bin/image-self-copy-blit/main.rs
@@ -48,7 +48,7 @@ use vulkano::{
     shader::PipelineShaderStageCreateInfo,
     swapchain::{
         acquire_next_image, AcquireError, Surface, Swapchain, SwapchainCreateInfo,
-        SwapchainCreationError, SwapchainPresentInfo,
+        SwapchainPresentInfo,
     },
     sync::{self, FlushError, GpuFuture},
     VulkanLibrary,
@@ -421,22 +421,21 @@ fn main() {
             recreate_swapchain = true;
         }
         Event::RedrawEventsCleared => {
-            let dimensions = window.inner_size();
-            if dimensions.width == 0 || dimensions.height == 0 {
-                return;
-            }
-
             previous_frame_end.as_mut().unwrap().cleanup_finished();
 
             if recreate_swapchain {
-                let (new_swapchain, new_images) = match swapchain.recreate(SwapchainCreateInfo {
-                    image_extent: dimensions.into(),
-                    ..swapchain.create_info()
-                }) {
-                    Ok(r) => r,
-                    Err(SwapchainCreationError::ImageExtentNotSupported { .. }) => return,
-                    Err(e) => panic!("failed to recreate swapchain: {e}"),
-                };
+                let image_extent: [u32; 2] = window.inner_size().into();
+
+                if image_extent.contains(&0) {
+                    return;
+                }
+
+                let (new_swapchain, new_images) = swapchain
+                    .recreate(SwapchainCreateInfo {
+                        image_extent,
+                        ..swapchain.create_info()
+                    })
+                    .expect("failed to recreate swapchain");
 
                 swapchain = new_swapchain;
                 framebuffers =
