+++ vulkano/src/pipeline/graphics/viewport.rs
@@ -232,16 +232,13 @@ impl From<Viewport> for ash::vk::Viewport {
     }
 }
 
-/// State of a single scissor box.
-// FIXME: add a check:
-//      Evaluation of (offset.x + extent.width) must not cause a signed integer addition overflow
-//      Evaluation of (offset.y + extent.height) must not cause a signed integer addition overflow
-#[derive(Debug, Copy, Clone, PartialEq, Eq)]
+/// A two-dimensional subregion.
+#[derive(Clone, Copy, Debug, PartialEq, Eq)]
 pub struct Scissor {
-    /// Coordinates in pixels of the top-left hand corner of the box.
+    /// Coordinates of the top-left hand corner of the box.
     pub origin: [u32; 2],
 
-    /// Dimensions in pixels of the box.
+    /// Dimensions of the box.
     pub dimensions: [u32; 2],
 }
 
@@ -276,6 +273,16 @@ impl From<Scissor> for ash::vk::Rect2D {
                 width: val.dimensions[0],
                 height: val.dimensions[1],
             },
+        }
+    }
+}
+
+impl From<ash::vk::Rect2D> for Scissor {
+    #[inline]
+    fn from(val: ash::vk::Rect2D) -> Self {
+        Scissor {
+            origin: [val.offset.x as u32, val.offset.y as u32],
+            dimensions: [val.extent.width, val.extent.height],
         }
     }
 }
